% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdist.R
\name{calc_aic}
\alias{calc_aic}
\title{Calculates the Akaike information criterion for the loglikelihood of a two
parameter probability distribution}
\usage{
calc_aic(loglik)
}
\arguments{
\item{loglik}{A vector or single number the loglikelihood of the model}
}
\value{
A single or vector of numerics equal to the input vector length
}
\description{
Calculates the Akaike information criterion for the loglikelihood of a two
parameter probability distribution
}
\examples{
calc_aic(loglik = -110)
}
