% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdist.R
\name{fit_cdt_dist}
\alias{fit_cdt_dist}
\title{Fits probability distributions using `coarseDataTools::dic.fit()`}
\usage{
fit_cdt_dist(data, models)
}
\arguments{
\item{data}{A numeric vector of values}

\item{models}{A character string or vector of character strings
specifying the names of the candidate models. This follows the R naming
convention for distributions, the density function is `d[name]`.}
}
\value{
A data frame of all models
}
\description{
Fits probability distributions using `coarseDataTools::dic.fit()`
}
\examples{
library(coarseDataTools)
data("nycH1N1")
fit_cdt_dist(data = nycH1N1, models = c("lnorm", "weibull"))
}
