% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdist.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{Calculates the goodness-of-fit statistics for the fitted distributions}
\usage{
goodness_of_fit(data, models)
}
\arguments{
\item{data}{A numeric vector of values}

\item{models}{A character string or vector of character strings
specifying the names of the candidate models. This follows the R naming
convention for distributions, the density function is `d[name]`.}
}
\value{
A `gofstat.fitdist` object from the `{fitdistrplus}` package
}
\description{
Calculates the goodness-of-fit statistics for the fitted distributions
}
\examples{
goodness_of_fit(
  data = rgamma(n = 100, shape = 1, scale = 1),
  models = c('gamma', 'weibull', 'lnorm')
)
}
