% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdist.R
\name{multi_fitdist}
\alias{multi_fitdist}
\title{Maximum likelihood fitting of multiple distributions}
\usage{
multi_fitdist(data, models)
}
\arguments{
\item{data}{A numeric vector of values}

\item{models}{A character string or vector of character strings
specifying the names of the candidate models. This follows the R naming
convention for distributions, the density function is `d[name]`.}
}
\value{
A data frame of all models
}
\description{
Maximum likelihood fitting of multiple distributions
}
\examples{
multi_fitdist(
  data = rgamma(n = 100, shape = 1, scale = 1),
  models = c("gamma", "weibull", "lnorm")
)
}
