% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdist.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{Calculates the goodness-of-fit statistics for the fitted distributions}
\usage{
goodness_of_fit(data, models, method = "mle")
}
\arguments{
\item{data}{A numeric vector of values}

\item{models}{A character string or vector of character strings
specifying the names of the candidate models. This follows the R naming
convention for distributions, the density function is `d[name]`.}

\item{method}{A character string for the method of model fitting.
See `?fitdistrplus::fitdist` for details.}
}
\value{
A `gofstat.fitdist` object from the `{fitdistrplus}` package
}
\description{
Calculates the goodness-of-fit statistics for the fitted distributions
}
\examples{
stub
}
