% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdist.R
\name{multi_fitdist}
\alias{multi_fitdist}
\title{Fits multiple distributions}
\usage{
multi_fitdist(data, models, method = "mle")
}
\arguments{
\item{data}{A numeric vector of values}

\item{models}{A character string or vector of character strings
specifying the names of the candidate models. This follows the R naming
convention for distributions, the density function is `d[name]`.}

\item{method}{A character string for the method of model fitting.
See `?fitdistrplus::fitdist` for details.}
}
\value{
A data frame of all models
}
\description{
Fits multiple distributions
}
\examples{
stub
}
